refactor: mineField 뷰 데이터 변환 책임 분리
